No need for MSF to do anything - just ask underlying. Put default implementation on Storage itself. Move computation to ObjectStorage.
