+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/BoolStorage.java
@@ -61,11 +61,6 @@ public final class BoolStorage extends Storage<Boolean> {
   }
 
   @Override
-  public StorageType getValuesType() {
-    return getType();
-  }
-
-  @Override
   public Boolean getItemBoxed(int idx) {
     return isMissing.get(idx) ? null : getItem(idx);
   }
