+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/SpecializedStorage.java
@@ -18,11 +18,6 @@ public abstract class SpecializedStorage<T> extends Storage<T> {
   @Override
   public abstract StorageType getType();
 
-  @Override
-  public StorageType getValuesType() {
-    return getType();
-  }
-
   /**
    * @param data the underlying data
    * @param size the number of items stored
